Reso omogenei i return type della funzione di query

+++ data/queries/deleteProducts.js
@@ -1,17 +1,23 @@
 import connection from "../db.js"
 
 async function deleteProducts(slug) {
+    
+    const foundProduct = request.product;
+    if(!foundProduct){
+        return {result:null, error:404};
+    }
 
     const deleteQuery = `
     DELETE FROM products WHERE slug = ?
     `;
 
+
     try {
-        const results = await connection.execute(deleteQuery, [slug]);
-        return results;
+        const result = await connection.execute(deleteQuery, [slug]);
+        return {result: result[0], error:null};
     }
     catch (error) {
-        return null;
+        return {result:null, error:500};
     }
 }
 export default deleteProducts
